+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 3.1 Beta 2 (16.01.2021)
+
+### Bug Fixes
+
+-  Fixed Gifu possibly stalling, if hand throttle hasn't been adjusted
+
 ## 3.1 Beta 1 (15.01.2021)
 
 ### Added